Move nanoc content into www
If the repository doesn't contain a `www` director, the static buildpack
will create one and move some of the files in the repo's root directory
into it.  As we're using multiple buildpacks, this logic is not correct
for us.

Adding the `www` directory ourselves for the nanoc related content
allows us to control exactly what is in there.